Move each loss to separate files
